# ghayath/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: "GHAYATH PERSONAL AI - Core API"
  version: "1.1.0"
  description: |
    REST/JSON API contract for **GHAYATH PERSONAL AI** (v1.1) — a Modular Agentic
    Architecture where an autonomous agent monitors and operates the owner's projects,
    inboxes, and communications under strict permission + approval control.

    ## Global rules

    **Envelope.** Every response (success or error) uses exactly the standard envelope:

    ```json
    { "success": true,  "data": { }, "request_id": "req_01J...", "timestamp": "2026-09-11T20:00:00Z" }
    { "success": false, "error": { "code": "PERMISSION_DENIED", "message": "...", "details": { } },
      "request_id": "req_01J...", "timestamp": "2026-09-11T20:00:00Z" }
    ```

    **Time.** All timestamps are ISO-8601 UTC (`...Z`) internally. Clients convert to
    the display timezone. Dates use `YYYY-MM-DD`.

    **IDs.** String identifiers with domain prefixes (ULID-based, human-greppable,
    log-safe): `usr_`, `prj_`, `ts_`, `mem_`, `cmd_`, `plan_`, `exec_`, `appr_`,
    `auto_`, `notf_`, `inc_`, `conv_`, `msg_` (email), `wmsg_` (WhatsApp),
    `draft_`, `repo_`, `iss_`, `evt_`, `req_`, `chk_`.

    **Authentication.** `Authorization: Bearer <JWT>` on every operation except
    `POST /auth/login`, `GET /health`, and `POST /webhooks/whatsapp`
    (signed via `X-Webhook-Signature`, HMAC-SHA256 of the raw body).
    Roles (RBAC): `OWNER` (full control), `AGENT` (service principal of the internal
...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /events/stream:
    get:
      tags:
      - Events
      operationId: getEventsStream
      summary: Server-Sent Events stream (realtime dashboard feed)
      description: "SSE from the `events` table. Cursor = last event ULID (`since`).\n\
        Heartbeat comment every 15s. Reconnect with `since=<last id>` → no loss,\nno duplicates.\n\
        AUTH (documented exception — the only path allowing a query token):\n  - `Authorization:\
        \ Bearer <token>` (preferred), or\n  - `?access_token=<token>` (EventSource compatibility;\
        \ use short-lived\n    access tokens — the token is visible in the URL).\nRoles:\
        \ OWNER, VIEWER. (AGENT excluded — least privilege.)\n"
      parameters:
      - name: since
        in: query
        required: false
        description: Event ULID cursor; omitted = start from now
        schema:
          type: string
      - name: types
        in: query
        required: false
        description: Comma-separated event names (from v1 events.event enum)
        schema:
          type: string
      - name: access_token
        in: query
        required: false
        description: 'PROPOSAL exception: bearer token via query for EventSource'
        schema:
          type: string
      responses:
        '200':
          description: Event stream
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /incidents/{incident_id}:
    patch:
      tags:
      - Incidents
      operationId: patchIncident
      summary: Transition incident status
      description: "OWNER, AGENT. Allowed transitions (else 409 CONFLICT):\n  OPEN→INVESTIGATING|RESOLVED;\
        \ INVESTIGATING→MITIGATED|RESOLVED;\n  MITIGATED→RESOLVED|INVESTIGATING; RESOLVED→CLOSED;\
        \ CLOSED→(terminal).\nNo-op (same status) → 409. `resolved_at` auto-set on RESOLVED.\n\
        `resolution_note` is recorded ONLY in audit_logs.details (no DDL change).\nIdempotency-Key\
        \ REQUIRED (24h TTL, v1 semantics).\n"
      parameters:
      - name: incident_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentStatusUpdateRequest'
      responses:
        '200':
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeIncident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Transition not allowed per the state matrix
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                conflict:
                  value:
                    success: false
                    error:
                      code: CONFLICT
                      message: Transition CLOSED→INVESTIGATING is not allowed
                      details: {}
                    request_id: req_01J9ZK7Q3M4N5P6R7S8T9V
                    timestamp: '2026-09-11T20:00:05Z'

  # ─────────────────────────── AUTOMATIONS ───────────────────────────
  /automations:
    get:
      tags:
      - Automations
      operationId: listAutomations
      summary: List automations
      description: 'Roles: OWNER, AGENT, VIEWER (read-only). No pagination (v1 list style).'
      parameters:
      - name: enabled
        in: query
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Automation list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeAutomationList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Automations]
      operationId: createAutomation
      summary: Create an automation
      description: |
        Registers a scheduled/event-driven automation. The scheduler persists
        `next_run_at`; each run is audit-logged and emits `automation.triggered`.
        `SCHEDULE` triggers require a 5-field cron; `EVENT` triggers bind to an
        internal event-bus name.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAutomationRequest'
            example:
              name: Monitor SANAA
              trigger:
                type: SCHEDULE
                cron: "*/15 * * * *"
              action:
                type: PROJECT_MONITOR
                target: prj_sanaa
              enabled: true
      responses:
        '201':
          description: Automation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /automations/{automation_id}:
    delete:
      tags:
      - Automations
      operationId: deleteAutomation
      summary: Delete an automation (hard delete, audited)
      description: 'OWNER only. Hard DELETE + immutable audit row (action DELETE_AUTOMATION).

        Idempotency-Key REQUIRED (24h TTL, v1 semantics). No approval (documented

        decision — not a sensitive tool operation).

        '
      parameters:
      - name: automation_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/IdempotencyKeyHeader'
      responses:
        '204':
          description: Deleted (or idempotent replay of a previous delete)
        '400':
          description: Missing/invalid Idempotency-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Automations]
      operationId: updateAutomation
      summary: Update an automation
      description: "Partial update. Disabling (enabled: false) keeps state; re-enabling recomputes next_run_at."
      parameters:
        - $ref: '#/components/parameters/AutomationIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAutomationRequest'
            example:
              enabled: false
      responses:
        '200':
          description: Automation updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ─────────────────────────── APPROVALS ───────────────────────────
...
  /incidents:
    post:
      tags: [Incidents]
      operationId: createIncident
      summary: Open an emergency incident
      description: |
        Raises a CRITICAL-path incident: immediate notification routing
        (PUSH + WHATSAPP), an `incident.created` event, and the agent's
        incident-response workflow. Idempotent per `Idempotency-Key`
        (a retry never opens a second incident).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncidentRequest'
            example:
              severity: CRITICAL
              system: SANAA
              issue: Database unavailable
              impact: Production unavailable
      responses:
        '201':
          description: Incident opened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentResponse'
              example:
                success: true
                data:
                  incident_id: inc_001
                  status: OPEN
                  severity: CRITICAL
                  system: SANAA
                  created_at: "2026-09-11T20:10:00Z"
                request_id: req_01J9ZK7Q3M4N5P6R7S8T9V
                timestamp: "2026-09-11T20:10:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ─────────────────────────── SYSTEM ───────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by `POST /auth/login` (1h) — refresh via `POST /auth/refresh`.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: HMAC-SHA256 hex digest of the raw request body, keyed with the provider's shared webhook secret.
  parameters:
...
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated unique key (UUID or ULID recommended). Retries carrying the
        same key receive the original response without re-executing the operation —
        this is what prevents double sends and double executions. Keys are retained 24h.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._-]{8,128}$'
        example: 7c9a4f2e-6b1d-4e8a-9c3f-1a2b3c4d5e6f
...
    AutomationIdParam:
      name: automation_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^auto_[a-z0-9]+$'
        example: auto_001
  headers:
...
    XRateLimitReset:
      description: Unix epoch seconds at which the rate-limit window resets.
      schema:
        type: integer
        minimum: 0
  responses:
    BadRequest:
      description: Malformed request or failed validation (`VALIDATION_ERROR`; field problems in `error.details.fields`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing/invalid/expired credentials (`AUTHENTICATION_FAILED` or `INVALID_TOKEN`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not permitted (`PERMISSION_DENIED`; required permission in `error.details.required`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource does not exist (`RESOURCE_NOT_FOUND`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Rate limit exceeded (`RATE_LIMITED`).
      headers:
        X-RateLimit-Reset:
          $ref: '#/components/headers/XRateLimitReset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: RATE_LIMITED
              message: "Rate limit exceeded, retry after the reset"
              details: {}
            request_id: req_01J9ZK7Q3M4N5P6R7S8T9V
            timestamp: "2026-09-11T20:00:00Z"
...
  schemas:
...
    SuccessEnvelope:
      type: object
      description: Base of every successful response.
      required: [success, request_id, timestamp]
      properties:
        success:
          type: boolean
          enum: [true]
        request_id:
          type: string
          description: Unique correlation id assigned by the gateway.
          example: req_01J9ZK7Q3M4N5P6R7S8T9V
        timestamp:
          type: string
          format: date-time
          description: Server response time, ISO-8601 UTC.
          example: "2026-09-11T20:00:00Z"
    ErrorCode:
      type: string
      description: Fixed error-code set. Do not introduce new codes without a spec revision.
      enum:
        - AUTHENTICATION_FAILED
        - INVALID_TOKEN
        - PERMISSION_DENIED
        - APPROVAL_REQUIRED
        - VALIDATION_ERROR
        - RESOURCE_NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - INTEGRATION_OFFLINE
        - TOOL_UNAVAILABLE
        - EXECUTION_FAILED
        - VERIFICATION_FAILED
        - TIMEOUT
        - DEPENDENCY_FAILURE
        - SECRET_ACCESS_DENIED
    ErrorResponse:
      type: object
      description: Base of every error response.
      required: [success, error, request_id, timestamp]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: Human-readable message (may be localized).
            details:
              type: object
              additionalProperties: true
              description: Structured context — field errors, approval_id, required permission, policy name, ...
        request_id:
          type: string
          example: req_01J9ZK7Q3M4N5P6R7S8T9V
        timestamp:
          type: string
          format: date-time
          example: "2026-09-11T20:00:00Z"

    # ─────────── Shared enums ───────────
...
    TriggerType:
      type: string
      enum: [SCHEDULE, EVENT, MANUAL]
    Trigger:
      type: object
      required: [type]
      properties:
        type:
          $ref: '#/components/schemas/TriggerType'
        cron:
          type: string
          description: 5-field cron expression (UTC). Required when type = SCHEDULE.
          example: "*/15 * * * *"
        event:
          type: string
          description: Internal event-bus name. Required when type = EVENT.
          example: deployment.failed
        interval_seconds:
          type: integer
          minimum: 60
          description: Optional fixed interval alternative for SCHEDULE.
    ActionType:
      type: string
      description: Known v1 action types.
      enum: [PROJECT_MONITOR, RUN_CHECK, CREATE_TASK, SEND_NOTIFICATION, SEND_EMAIL_DRAFT]
    Action:
      type: object
      required: [type]
      properties:
        type:
          $ref: '#/components/schemas/ActionType'
        target:
          type: string
          description: Target id (project, repository, ...).
          example: prj_sanaa
        params:
          type: object
          additionalProperties: true
          description: Action-specific parameters.
    Automation:
      type: object
      required: [id, name, trigger, action, enabled, created_at, updated_at]
      properties:
        id:
          type: string
          pattern: '^auto_[a-z0-9]+$'
          example: auto_001
        name:
          type: string
          minLength: 1
          maxLength: 160
          example: Monitor SANAA
        description:
          anyOf:
            - type: string
            - type: 'null'
        trigger:
          $ref: '#/components/schemas/Trigger'
        action:
          $ref: '#/components/schemas/Action'
        enabled:
          type: boolean
        last_run_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        last_run_status:
          anyOf:
            - type: string
              enum: [SUCCESS, FAILURE, SKIPPED]
            - type: 'null'
        next_run_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAutomationRequest:
      type: object
      required: [name, trigger, action]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: string
          maxLength: 500
        trigger:
          $ref: '#/components/schemas/Trigger'
        action:
          $ref: '#/components/schemas/Action'
        enabled:
          type: boolean
          default: true
    UpdateAutomationRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: string
          maxLength: 500
        trigger:
          $ref: '#/components/schemas/Trigger'
        action:
          $ref: '#/components/schemas/Action'
        enabled:
          type: boolean
    AutomationResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/Automation'

    # ─────────── Approvals ───────────
...
    IncidentSeverity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    CreateIncidentRequest:
      type: object
      required: [severity, system, issue, impact]
      properties:
        severity:
          $ref: '#/components/schemas/IncidentSeverity'
        system:
          type: string
          minLength: 1
          example: SANAA
        issue:
          type: string
          minLength: 1
          example: Database unavailable
        impact:
          type: string
          minLength: 1
          example: Production unavailable
        source:
          type: string
          enum: [agent, user, monitoring, integration]
          default: agent
    CreateIncidentData:
      type: object
      required: [incident_id, status, severity, system, created_at]
      properties:
        incident_id:
          type: string
          pattern: '^inc_[a-z0-9]+$'
          example: inc_001
        status:
          type: string
          enum: [OPEN]
        severity:
          $ref: '#/components/schemas/IncidentSeverity'
        system:
          type: string
        created_at:
          type: string
          format: date-time
    IncidentResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/CreateIncidentData'

    # ─────────── System ───────────
...
    IncidentStatusUpdateRequest:
      type: object
      required:
      - status
      properties:
        status:
          type: string
          enum:
          - OPEN
          - INVESTIGATING
          - MITIGATED
          - RESOLVED
          - CLOSED
        resolution_note:
          type: string
          maxLength: 2000
          description: Recorded in audit_logs.details only (no incident column)
...
    AutomationView:
      type: object
      required:
      - id
      - name
      - enabled
      - created_at
      - updated_at
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type:
          - string
          - 'null'
        trigger:
          type: object
        action:
          type: object
        enabled:
          type: boolean
        last_run_at:
          type:
          - string
          - 'null'
          format: date-time
        last_run_status:
          type:
          - string
          - 'null'
          enum:
          - SUCCESS
          - FAILURE
          - SKIPPED
          - null
        next_run_at:
          type:
          - string
          - 'null'
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EnvelopeAutomationList:
      allOf:
      - $ref: '#/components/schemas/SuccessEnvelope'
      - type: object
        properties:
          data:
            type: array
            items:
              $ref: '#/components/schemas/AutomationView'
    IncidentView:
      type: object
      required:
      - id
      - severity
      - status
      - created_at
      properties:
        id:
          type: string
        severity:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - CRITICAL
        system:
          type: string
        issue:
          type: string
        impact:
          type:
          - string
          - 'null'
        status:
          type: string
          enum:
          - OPEN
          - INVESTIGATING
          - MITIGATED
          - RESOLVED
          - CLOSED
        source:
          type: string
          enum:
          - agent
          - user
          - monitoring
          - integration
        resolved_at:
          type:
          - string
          - 'null'
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EnvelopeIncident:
      allOf:
      - $ref: '#/components/schemas/SuccessEnvelope'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/IncidentView'
...